yapyseq/requirements.txt
pyyaml
yamale